+++ Core/Executor/RepositoryExecutor.php
@@ -106,7 +106,7 @@ abstract class RepositoryExecutor extends AbstractExecutor implements LanguageAw
 
             $previousUserId = $this->loginUser(self::ADMIN_USER_ID);
             try {
-                $this->$action();
+                $output = $this->$action();
             } catch (\Exception $e) {
                 $this->loginUser($previousUserId);
                 throw $e;
@@ -115,6 +115,7 @@ abstract class RepositoryExecutor extends AbstractExecutor implements LanguageAw
             // reset the environment as much as possible as we had found it before the migration
             $this->loginUser($previousUserId);
 
+            return $output;
         } else {
             throw new \Exception("Invalid step definition: value '$action' is not a method of " . get_class($this));
         }